requirements.txt
Appium_Python_Client==1.3.0
Requests==2.31.0
